news_scrapper_service/requirements.txt
fastapi[standard]
fastapi_pagination
pydantic
sqlmodel
uvicorn
requests
python-dotenv
apscheduler
bs4
transformers
torch
aiohttp
